+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/listener/JobLevelListener.java
@@ -5,12 +5,19 @@ import com.asml.apa.wta.core.model.Domain;
 import com.asml.apa.wta.core.model.Task;
 import com.asml.apa.wta.core.model.Workflow;
 import com.asml.apa.wta.core.streams.Stream;
+import com.asml.apa.wta.core.model.enums.Domain;
+import com.asml.apa.wta.spark.dagsolver.DagSolver;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Stream;
 import lombok.Getter;
 import org.apache.spark.SparkContext;
 import org.apache.spark.scheduler.SparkListenerJobEnd;
 import org.apache.spark.scheduler.SparkListenerJobStart;
+import scala.collection.JavaConverters;
 
 /**
  * This class is a job-level listener for the Spark data source.
@@ -27,7 +34,7 @@ public class JobLevelListener extends AbstractListener<Workflow> {
 
   private final Map<Long, Long> jobSubmitTimes = new ConcurrentHashMap<>();
 
-  private final Map<Long, Integer> criticalPathTasks = new ConcurrentHashMap<>();
+  private final Map<Long, List<Task>> jobToStages = new ConcurrentHashMap<>();
 
   /**
    * Constructor for the job-level listener.
@@ -68,15 +75,27 @@ public class JobLevelListener extends AbstractListener<Workflow> {
 
   /**
    * Callback for job start event, tracks the submit time of the job.
+   * Also tracks the stages and their parents in the job.
    *
    * @param jobStart The SparkListenerJobStart event object containing information upon job start.
    * @author Henry Page
+   * @author Tianchen Qu
    * @since 1.0.0
    */
   @Override
   public void onJobStart(SparkListenerJobStart jobStart) {
     jobSubmitTimes.put(jobStart.jobId() + 1L, jobStart.time());
-    criticalPathTasks.put(jobStart.jobId() + 1L, jobStart.stageIds().length());
+    List<Task> stages = new ArrayList<>();
+    JavaConverters.seqAsJavaList(jobStart.stageInfos()).forEach(stageInfo -> {
+      stages.add(Task.builder()
+          .id((long) stageInfo.stageId() + 1)
+          .parents(JavaConverters.seqAsJavaList(stageInfo.parentIds()).stream()
+              .mapToInt(parentId -> (int) parentId + 1)
+              .mapToLong(parentId -> (long) parentId)
+              .toArray())
+          .build());
+    });
+    jobToStages.put((long) jobStart.jobId() + 1, stages);
   }
 
   /**
@@ -97,7 +116,7 @@ public class JobLevelListener extends AbstractListener<Workflow> {
     final long tsSubmit = jobSubmitTimes.get(jobId);
     final Stream<Task> tasks = wtaTaskListener.getWorkflowsToTasks().onKey(jobId);
     final long criticalPathLength = -1L;
-    final int criticalPathTaskCount = criticalPathTasks.get(jobId);
+    int criticalPathTaskCount = -1;
     final int maxNumberOfConcurrentTasks = -1;
     final String nfrs = "";
 
@@ -115,6 +134,39 @@ public class JobLevelListener extends AbstractListener<Workflow> {
         .filter(resourceAmount -> resourceAmount >= 0.0)
         .reduce(Double::sum)
         .orElse(-1.0);
+
+    if (config.isStageLevel()) {
+      stageLevelListener.setStages(jobId);
+    } else {
+      TaskLevelListener taskLevelListener = (TaskLevelListener) wtaTaskListener;
+      taskLevelListener.setTasks(stageLevelListener, jobId);
+
+      List<Task> jobStages = stageLevelListener.getProcessedObjects().stream()
+          .filter(stage -> stage.getWorkflowId() == jobId)
+          .collect(Collectors.toList());
+      jobStages.addAll(jobToStages.get(jobId).stream()
+          .filter(stage -> !jobStages.stream()
+              .map(Task::getId)
+              .collect(Collectors.toList())
+              .contains(stage.getId()))
+          .collect(Collectors.toList()));
+
+      final List<Task> criticalPath = solveCriticalPath(jobStages);
+      TaskLevelListener listener = (TaskLevelListener) wtaTaskListener;
+
+      final Map<Long, List<Task>> stageToTasks = listener.getStageToTasks();
+      criticalPathTaskCount = criticalPath.stream()
+          .map(stage -> stageToTasks.get(stage.getId()).size())
+          .reduce(Integer::sum)
+          .orElse(-1);
+      criticalPathLength = criticalPath.stream()
+          .map(stage -> stageToTasks.getOrDefault(stage.getId(), new ArrayList<>()).stream()
+              .map(Task::getRuntime)
+              .reduce(Long::max)
+              .orElse(0L))
+          .reduce(Long::sum)
+          .orElse(-1L);
+    }
 
     wtaTaskListener.removesWorkflowAssociations(jobId);
 
@@ -162,7 +214,6 @@ public class JobLevelListener extends AbstractListener<Workflow> {
   private void cleanUpContainers(SparkListenerJobEnd jobEnd) {
     long jobId = jobEnd.jobId() + 1;
     jobSubmitTimes.remove(jobId);
-    criticalPathTasks.remove(jobId);
 
     Stream<Long> stageIds = new Stream<>();
 
@@ -198,5 +249,44 @@ public class JobLevelListener extends AbstractListener<Workflow> {
    */
   private double computeSum(Stream<Double> data) {
     return data.filter(task -> task >= 0.0).reduce(Double::sum).orElse(-1.0);
+  }
+
+  /**
+   * This is a method called on application end. it sets up the resources used in the spark workflow.
+   * It also calculated the critical path for the job.
+   *
+   * @author Tianchen Qu
+   * @author Pil Kyu Cho
+   * @since 1.0.0
+   */
+  public void setWorkflows() {
+    final List<Workflow> workflows = this.getProcessedObjects();
+    workflows.forEach(workflow -> {
+      workflow.setTotalResources(Arrays.stream(workflow.getTasks())
+          .map(Task::getResourceAmountRequested)
+          .filter(resourceAmount -> resourceAmount >= 0.0)
+          .reduce(Double::sum)
+          .orElse(-1.0));
+    });
+  }
+
+  /**
+   * This method takes the stages inside this job and return the critical path.
+   * it will filter out all dummy caches nodes.
+   *
+   * @param stages all stages in the job(including the cached ones)
+   * @return critical path
+   * @author Tianchen Qu
+   * @since 1.0.0
+   */
+  List<Task> solveCriticalPath(List<Task> stages) {
+    try {
+      DagSolver dag = new DagSolver(stages, (TaskLevelListener) wtaTaskListener);
+      return dag.longestPath().stream()
+          .filter(stage -> stage.getRuntime() != 0)
+          .collect(Collectors.toList());
+    } catch (Exception e) {
+      return new ArrayList<>();
+    }
   }
 }