+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/EndToEnd.java
@@ -147,14 +147,16 @@ public class EndToEnd {
         .set("spark.plugins", "com.asml.apa.wta.spark.WtaPlugin")
         .set("spark.executor.instances", "2")
         .set("spark.executor.cores", "2")
+        .set("spark.logConf", "true")
         .set("spark.driver.extraJavaOptions", "-DconfigFile=" + args[0]);
     //        .set("spark.driver.extraJavaOptions", "-DconfigFile=" +
-    //                "adapter/spark/src/test/resources/config-stage.json");
+    // "adapter/spark/src/test/resources/config.json");
     SparkSession spark = SparkSession.builder().config(conf).getOrCreate();
     SparkContext sc = spark.sparkContext();
+    sc.setLogLevel("INFO");
     sparkOperation(JavaSparkContext.fromSparkContext(sc).textFile(args[1]));
-    //  sparkOperation(
-    //      JavaSparkContext.fromSparkContext(sc).textFile("adapter/spark/src/test/resources/e2e-input.txt"));
+    // sparkOperation(
+    //    JavaSparkContext.fromSparkContext(sc).textFile("adapter/spark/src/test/resources/e2e-input.txt"));
     sc.stop();
   }
 }