+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/driver/WtaDriverPluginTest.java
@@ -105,7 +105,8 @@ class WtaDriverPluginTest {
     } catch (Exception ignored) {
     }
 
-    verify(mockedSparkContext, times(0))
+    verify(sut, times(1)).removeListeners();
+    verify(mockedSparkContext, times(1))
         .removeSparkListener(sut.getSparkDataSource().getTaskLevelListener());
     verify(mockedSparkContext, times(1))
         .removeSparkListener(sut.getSparkDataSource().getStageLevelListener());