+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/listener/JobLevelListenerTest.java
@@ -248,6 +248,8 @@ class JobLevelListenerTest extends BaseLevelListenerTest {
     ListBuffer<StageInfo> stageBuffer2 = new ListBuffer<>();
     stageBuffer1.$plus$eq(testStageInfo1);
     stageBuffer1.$plus$eq(testStageInfo2);
+    stageBuffer2.$plus$eq(testStageInfo1);
+    stageBuffer2.$plus$eq(testStageInfo2);
     stageBuffer2.$plus$eq(testStageInfo3);
     stageBuffer2.$plus$eq(testStageInfo4);
     SparkListenerJobStart jobStart1 =
@@ -287,6 +289,10 @@ class JobLevelListenerTest extends BaseLevelListenerTest {
     fakeTaskListener1.onJobStart(jobStart2);
     fakeStageListener1.onJobStart(jobStart2);
     fakeJobListener1.onJobStart(jobStart2);
+    fakeTaskListener1.onTaskEnd(taskEndEvent1);
+    fakeStageListener1.onStageCompleted(stageCompleted1);
+    fakeTaskListener1.onTaskEnd(taskEndEvent2);
+    fakeStageListener1.onStageCompleted(stageCompleted2);
     fakeTaskListener1.onTaskEnd(taskEndEvent3);
     fakeStageListener1.onStageCompleted(stageCompleted3);
     fakeTaskListener1.onTaskEnd(taskEndEvent4);
@@ -302,6 +308,7 @@ class JobLevelListenerTest extends BaseLevelListenerTest {
     Task task4 = fakeTaskListener1.getProcessedObjects().drop(3).head();
     assertThat(task4.getParents().length).isEqualTo(1);
     assertThat(task4.getParents()).contains(taskId3 + 1);
+    assertThat(task4.getParents()).contains(taskId2 + 1);
     assertThat(task4.getChildren().length).isEqualTo(0);
   }