+++ core/src/main/java/com/asml/apa/wta/core/model/ResourceState.java
@@ -66,6 +66,7 @@ public class ResourceState implements BaseTraceObject {
    */
   @Override
   public GenericRecord convertToRecord(ParquetSchema schema) {
+    log.trace("Converting ResourceState with Resource id to record");
     return schema.convertFromPojo(this, ResourceState.class);
   }
 }