+++ core/src/main/java/com/asml/apa/wta/core/model/Task.java
@@ -3,6 +3,7 @@ package com.asml.apa.wta.core.model;
 import com.asml.apa.wta.core.io.ParquetSchema;
 import lombok.Builder;
 import lombok.Data;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.avro.generic.GenericRecord;
 
 /**
@@ -13,6 +14,7 @@ import org.apache.avro.generic.GenericRecord;
  */
 @Data
 @Builder
+@Slf4j
 public class Task implements BaseTraceObject {
 
   private static final long serialVersionUID = -1372345471722101373L;
@@ -70,6 +72,7 @@ public class Task implements BaseTraceObject {
    */
   @Override
   public GenericRecord convertToRecord(ParquetSchema schema) {
+    log.trace("Converting Task with id {} to record", this.id);
     return schema.convertFromPojo(this, Task.class);
   }
 }