+++ core/src/main/java/com/asml/apa/wta/core/model/Workflow.java
@@ -3,6 +3,7 @@ package com.asml.apa.wta.core.model;
 import com.asml.apa.wta.core.io.ParquetSchema;
 import lombok.Builder;
 import lombok.Data;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.avro.generic.GenericRecord;
 
 /**
@@ -13,6 +14,7 @@ import org.apache.avro.generic.GenericRecord;
  */
 @Data
 @Builder
+@Slf4j
 public class Workflow implements BaseTraceObject {
 
   private static final long serialVersionUID = 9065743819019553490L;
@@ -25,7 +27,7 @@ public class Workflow implements BaseTraceObject {
 
   private final long taskCount;
 
-  private final long criticalPathLength;
+  private long criticalPathLength;
 
   private final long criticalPathTaskCount;
 
@@ -62,6 +64,7 @@ public class Workflow implements BaseTraceObject {
    */
   @Override
   public GenericRecord convertToRecord(ParquetSchema schema) {
+    log.trace("Converting Workflow with id {} to record", this.id);
     return schema.convertFromPojo(this, Workflow.class);
   }
 }