+++ core/src/main/java/com/asml/apa/wta/core/utils/ShellUtils.java
@@ -19,6 +19,7 @@ public class ShellUtils {
    * @since 1.0.0
    */
   public CompletableFuture<String> executeCommand(String command, boolean suppressErrors) {
+    log.trace("Executing shell command: {}", command);
     return CompletableFuture.supplyAsync(() -> {
       try {
         String[] commands = {"sh", "-c", command};
@@ -51,6 +52,7 @@ public class ShellUtils {
    * @since 1.0.0
    */
   private String readProcessOutput(Process process) {
+    log.trace("Reading shell command outputs.");
     try (BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()))) {
       StringBuilder output = new StringBuilder();
       String line;